fix: use exact ID delete in span promotion, use BigInteger for counts
- Change promotion DELETE from range-based (id >= min AND id <= max) to
  exact match (id = ANY(...)) to prevent over-deleting rows inserted by
  concurrent workers between SELECT and DELETE
- Change count and error_count from PositiveIntegerField (max ~2.1B) to
  PositiveBigIntegerField to prevent overflow on high-traffic endpoints
  that accumulate counts over their lifetime

+++ apps/performance/migrations/0021_performance_v3.py
@@ -113,12 +113,12 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name="transactiongroup",
             name="count",
-            field=models.PositiveIntegerField(default=0),
+            field=models.PositiveBigIntegerField(default=0),
         ),
         migrations.AddField(
             model_name="transactiongroup",
             name="error_count",
-            field=models.PositiveIntegerField(default=0),
+            field=models.PositiveBigIntegerField(default=0),
         ),
         migrations.AddField(
             model_name="transactiongroup",
